ref(ui): Consolidate server frontend hydration logic
 - Consolidates all frontend hydration data into a single global window
   variable `__initialData`.

 - Splits application bootstraping into `bootstrap.jsx`, to allow for
   initial data setup before bootstrapping the application.
