--- src/sentry/templates/sentry/includes/sdk-config.html
@@ -1,10 +1,0 @@
-{% load sentry_api %}
-{% load sentry_dsn %}
-<script>
-  window.__SENTRY__OPTIONS = {
-    dsn: '{% public_dsn %}',
-    release: '{{ sentry_version.build }}',
-    whitelistUrls: {% convert_to_json ALLOWED_HOSTS %}
-  };
-  window.__SENTRY__USER = {% get_user_context request %} || undefined;
-</script>
