+++ src/sentry/web/helpers.py
@@ -27,11 +27,9 @@ def get_default_context(request, existing_context=None, team=None):
     from sentry.plugins import plugins
 
     context = {
-        'CSRF_COOKIE_NAME': settings.CSRF_COOKIE_NAME,
         'URL_PREFIX': options.get('system.url-prefix'),
         'SINGLE_ORGANIZATION': settings.SENTRY_SINGLE_ORGANIZATION,
         'PLUGINS': plugins,
-        'ALLOWED_HOSTS': list(settings.ALLOWED_HOSTS),
         'ONPREMISE': settings.SENTRY_ONPREMISE,
     }
